requirements.txt
certifi==2018.8.13
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1
<<<<<<< HEAD
django-heroku==0.3.1
gunicorn==19.9.0
heroku==0.1.4
idna==2.7
=======
djangorestframework==3.8.2
>>>>>>> 3d6f411fb522fe72bfd09a16ab9d3eb9a3a34627
psycopg2==2.7.5
python-dateutil==1.5
python-decouple==3.1
pytz==2018.5
requests==2.19.1
urllib3==1.23
whitenoise==4.0
